Added multiple features.
- Fully dynamic stack creation.
- Fully dynamic stack deletion.
- Ability to reset the entire skill.

+++ cloud-former-lambda/Python/tempf/lambda_function.py
@@ -2,11 +2,11 @@
 #Alexa Python (2.7) Lambda Skill
 #Author: Jordan Lindsey
 #Email: jordan.lindsey@capgemini.com
-#Version: 3.3
+#Version: 3.4
 #Date: 18/07/2017
-#Changelog: Added basic conversations. Added 2FA timeout. Multi-user verification.
-#Features: Can give current date/time. Creation of AWS stack. Deletion of AWS stack. Conversations. SMS Verification.
-#In-Production: Requesting template information.
+#Changelog: Added basic conversations. Added 2FA timeout. Multi-user verification. Requesting available stacks.
+#Features: Can give current date/time. Creation of AWS stack. Deletion of AWS stack. Conversations. SMS Verification. Dynamic Stack Formation.
+#In-Production: Listing current status of deployed stacks.
 ##
 
 ##Begin function
@@ -22,6 +22,7 @@ from flask_ask import Ask, statement, question, session, convert_errors
 from flask import Flask, render_template
 from random import randint
 import csv
+import ast
 
 print('Loading function')
 
@@ -43,6 +44,25 @@ def cancel():
 @ask.session_ended
 def session_ended():
     return "{}", 200
+
+@ask.intent("ResetSkill")
+def reset_skill():
+    s3 = boto3.client('s3')
+    open("/tmp/blank.txt","w").close()
+    with open('/tmp/blank.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'availabletemplates.txt')
+    with open('/tmp/blank.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'number.txt')
+    with open('/tmp/blank.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'request.txt')
+    with open('/tmp/blank.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'securitycode.txt')
+    with open('/tmp/blank.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'unknown.txt')
+    with open('/tmp/blank.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'user.txt')
+    speech_output = "Skill reset."
+    return statement(speech_output)
 
 @ask.intent("GetStatus")
 def get_system_status():
@@ -190,9 +210,6 @@ def delete_instance(number,code,user):
     else:
         return question("Incorrect code.")
 
-#@ask.intent("StackStatus") #in-development
-#def stack_status(name):
-
 @ask.intent("UnknownRequest")
 def unknown_request(number,code,user):
     if number == None and code == None and user == None:
@@ -203,10 +220,12 @@ def unknown_request(number,code,user):
     KEY = 'request.txt'
     KEY2 = 'number.txt'
     KEY3 = 'unknown.txt'
+    KEY4 = 'availabletemplates.txt'
     try:
         s3.Bucket(BUCKET_NAME).download_file(KEY, '/tmp/request.txt')
         s3.Bucket(BUCKET_NAME).download_file(KEY2, '/tmp/number.txt')
         s3.Bucket(BUCKET_NAME).download_file(KEY3, '/tmp/unknown.txt')
+        s3.Bucket(BUCKET_NAME).download_file(KEY4, '/tmp/availabletemplates.txt')
     except botocore.exceptions.ClientError as e:
         print("An error has occured.")
         if e.response['Error']['Code'] == "404":
@@ -224,11 +243,27 @@ def unknown_request(number,code,user):
     file3=open("/tmp/unknown.txt","r")
     unknownrequest=file3.read()
 
+    file4=open("/tmp/availabletemplates.txt","r")
+    availabletemplates=file4.read()
+
+    if availabletemplates == None or availabletemplates == "":
+        return question("Please request which templates are available before proceeding. This is to prevent modifying the incorrect stack.")
+
+    if number == None:
+        number = code
+    else:
+        pass
+
     s3 = boto3.client('s3')
     if unknownrequest == "coderequest":
         code=int(number)
         number=None
     elif unknownrequest == "numberrequest":
+        file=open("/tmp/number.txt","w")
+        file.write(number)
+        file.close()
+        with open('/tmp/number.txt', 'rb') as data:
+            s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'number.txt')
         number=int(number)
         code=None
     elif unknownrequest == "userrequest":
@@ -307,15 +342,34 @@ def unknown_request(number,code,user):
 
 def stackformation(number):
     client = boto3.client('cloudformation')
+    s3 = boto3.resource('s3')
+    BUCKET_NAME = 'jlindsey-bucket-eu-west-1'
+    KEY = 'availabletemplates.txt'
+    try:
+        s3.Bucket(BUCKET_NAME).download_file(KEY, '/tmp/availabletemplates.txt')
+    except botocore.exceptions.ClientError as e:
+        print("An error has occured.")
+        if e.response['Error']['Code'] == "404":
+            print("The object does not exist.")
+            return statement("An error has occured. Please check the Alexa configuration.")
+        else:
+            raise
+
+    file=open("/tmp/availabletemplates.txt","r")
+    liststring=file.read()
+    liststring2=ast.literal_eval(liststring)
+    if int(number)>len(liststring2) or number == None:
+        speech_output = "The number you specified is invalid."
+        return speech_output
     try:
         response = client.create_stack(
-            StackName='Cloud-Former',
-            TemplateURL='https://s3-eu-west-1.amazonaws.com/jlindsey-bucket-eu-west-1/basic_ec2_instance.json',
-            TimeoutInMinutes=2,
+            StackName='Cloud-Former-'+str(number),
+            TemplateURL='https://s3-eu-west-1.amazonaws.com/jlindsey-bucket-eu-west-1/'+str(liststring2[int(number)-1]),
+            TimeoutInMinutes=5,
             OnFailure='ROLLBACK',
-            ClientRequestToken='tokenrequest1'
+            ClientRequestToken='tokenrequest'+str(number)
         )
-        speech_output = "Your stack has been launched."#.format(number)
+        speech_output = "Stack "+number+" has been launched."
         print ("Success.")
     except Exception as e:
         print('Stack formation failed.')
@@ -324,18 +378,39 @@ def stackformation(number):
 
 def stackdeletion(number):
     client = boto3.client('cloudformation')
+    s3 = boto3.resource('s3')
+    BUCKET_NAME = 'jlindsey-bucket-eu-west-1'
+    KEY = 'availabletemplates.txt'
+    try:
+        s3.Bucket(BUCKET_NAME).download_file(KEY, '/tmp/availabletemplates.txt')
+    except botocore.exceptions.ClientError as e:
+        print("An error has occured.")
+        if e.response['Error']['Code'] == "404":
+            print("The object does not exist.")
+            return statement("An error has occured. Please check the Alexa configuration.")
+        else:
+            raise
+
+    file=open("/tmp/availabletemplates.txt","r")
+    liststring=file.read()
+    liststring2=ast.literal_eval(liststring)
+    if int(number)>len(liststring2) or number == None:
+        speech_output = "The number you specified is invalid."
+        return speech_output
     try:
         response = client.delete_stack(
-            StackName='Cloud-Former',
-            ClientRequestToken='tokenrequest2'
+            StackName='Cloud-Former-'+str(number),
+            ClientRequestToken='tokenrequest0'+str(number)
         )
-        speech_output = "Your stack has been deleted."
+        speech_output = "Stack "+number+" has been deleted."
     except Exception as e:
         print('Stack deletion failed.')
         speech_output = "There has been a problem. The instance was not deleted successfully."
+        raise e
     return speech_output
 
 def security_request(user):
+    #return True #enable for debugging - warning: this disables all security!
     s3 = boto3.resource('s3')
     BUCKET_NAME = 'jlindsey-bucket-eu-west-1'
     KEY = 'contacts.csv'
@@ -381,6 +456,7 @@ def security_request(user):
     return True
 
 def security_check(code):
+    #return True #enable for debugging - warning: this disables all security!
     print("Checking...")
     s3 = boto3.resource('s3')
     BUCKET_NAME = 'jlindsey-bucket-eu-west-1'
@@ -424,13 +500,40 @@ def request_list():
     speech_output = "Function not yet completed."
     return statement(speech_output)
 
-@ask.intent("ListStacks")
+@ask.intent("RequestInstance")
 def list_stacks():
     s3 = boto3.resource('s3')
-    my_bucket = s3.Bucket('some/path/')
+    BUCKET_NAME = s3.Bucket('jlindsey-bucket-eu-west-1')
+    list1=[]
+    list2=[]
+    i=0
+    for file in BUCKET_NAME.objects.all():
+        filename=file.key
+        words = filename.split(".")
+        title=words[0]
+        extension=words[1]
+        if extension == "json":
+            i+=1
+            list1.append(str(i)+'. '+title.replace("_", " "))
+            list2.append(file.key)
 
+    s3 = boto3.client('s3')
+    open("/tmp/availabletemplates.txt","w").close()
+    file=open("/tmp/availabletemplates.txt","w")
+    file.write(str(list2))
+    file.close()
+    with open('/tmp/availabletemplates.txt', 'rb') as data:
+        s3.upload_fileobj(data, 'jlindsey-bucket-eu-west-1', 'availabletemplates.txt')
 
+    speech_output = '. '.join(list1)
 
-    speech_output = "Function not yet completed."
+    #print(ast.literal_eval(str(list2)))
     return statement(speech_output)
+
+#@ask.intent("StackStatus") #in-development
+#def stack_status(name):
+
+#@ask.intent("TemplateSummary") #in-development
+#def template_summary(name):
+
 ###
